Move screenshot countdown logic from Keyboard.cpp to Screenshot.cpp
