+++ src/OpenLoco/src/Ui/Screenshot.h
@@ -5,6 +5,15 @@
 
 namespace OpenLoco::Input
 {
+    enum class ScreenshotType : uint8_t
+    {
+        regular = 0,
+        giant = 1,
+    };
+
+    void triggerScreenshotCountdown(int8_t numTicks, ScreenshotType type);
+    void handleScreenshotCountdown();
+
     std::string saveScreenshot();
     std::string saveGiantScreenshot();
 }
