+++ src/OpenLoco/src/Windows/ToolbarTopAlt.cpp
@@ -20,6 +20,7 @@
 #include "S5/S5.h"
 #include "ToolbarTopCommon.h"
 #include "Ui/Dropdown.h"
+#include "Ui/Screenshot.h"
 #include "Ui/WindowManager.h"
 #include "Vehicles/Vehicle.h"
 #include "Widget.h"
